python/requirements.txt
pyyaml
pymysql
cerberus
flask
celery
blinker
sqlalchemy
flask_sqlalchemy
flask-security
webcolors
requests